style: 컴포넌트 별 폴더 생성
