+++ src/components/common/Input/Input.tsx
@@ -1,0 +1,16 @@
+import classNames from 'classnames/bind'
+import styles from './Input.module.scss'
+import { Controller, RegisterOptions, useFormContext } from 'react-hook-form'
+
+const cx = classNames.bind(styles)
+
+interface InputProps {
+  size: 'lg' | 'sm'
+  type: 'text' | 'password'
+}
+
+const Input = ({ size, type }: InputProps) => {
+  return <input className={cx('input', `input-size-${size}`)} type={type} />
+}
+
+export default Input
